refactor: 과거 메시지 조회 로직 분리
